Fix wrong image volume causing segfault
When finding the set of blocks containing a bounding box, BlockUtils
used the floor of the min of the box and block size and the ceiling of
the extent of the box and block size, but the difference between the
floor and the min was not considered. In
SegmentGuarantor::computeFeatures this led to cases where the set of
blocks chosen to bound the set of segment slices did not cover the max
of their bounding box. Consequently the image volume retrieved by
StackStore was too small. Consequently
HistogramFeatureExtractor::computeHistogram would access, for some
slices, pixels outside the image volume. However, because there is no
bounds checking, this behavior went unnoticed until it began causing
segfaults do to unrelated changes in memory behavior elsewhere in the
program (the segment store).