Add explicit constraints to solution
Use explicit constraints from the stores, e.g., user corrections to the
solution or constraints from prior tracing, when computing solutions.
Add methods for retrieving these constraints on segments from the
SegmentStore. For now the constraints are always greater-or-equal
constraints with coefficients and values of one, though this will be
extended later.

Tested to verify no constraints do not affect solution, and constraints
requiring a segment from a slice otherwise not chosen by the solver do
cause that slice to be in the solution.