PostgreSqlSliceStore: fix conflict set block relation
95b76c94cd701c634e7427d97ef93add105a2e4d combined conflict set insertion
into a single query with a common table expression (CTE).
Sub-statements in the WITH clause of a CTE only execute when required
by the main statement (including via another WITH statement), or
otherwise concurrently with the main statement. Consequently the main
conflict set to block relation statement did not select all of the
conflict sets inserted by the WITH clause. The main statement could
reference the WITH INSERT statement to force execution order, except
that requires that the statement have a RETURNING clause (even if
unused), which it cannot because djsopnet_sliceconflictset has a
conditional DO INSTEAD NOTHING RULE which does not yet support
RETURNING in PostgreSQL. The RULE could be replaced with a trigger
(which would allow RETURNING to be used), but constraint exception
triggers are significantly slower than DO INSTEAD NOTHING rules.

Instead use two INSERTs (in a single query) for conflict set creation
and block relation. Because this requires transmitting and parsing the
set tuples twice, this is ~15% slower per effective insert but
guaranteed correct.