SliceGuarantor: fix lost conflict sets
When extracting slices, slices were being translated to account for
block coordinates while conflict sets were not. Because slice
translation changes the slice hash, this caused conflict sets for block
sets with an upper left bounder other than (0, 0) to not be collected for
containing slices in requested blocks and thus discarded.

Fix by recreating conflict sets with correct hashes after translation.
Greatly increases the number of stored conflict sets and reduces the
percentage of slices without conflicts from ~70% to ~15% in the test
dataset. Greatly reduces the number of segments as fewer slice
combinations are viable segments due to conflicts.