Read feature weights from segment store
Have SolutionGuarantor read feature weights from the SegmentStore.
Features are segment-related data, so until there is a project/solution
configuration store this is the most reasonable location (and it can be
easily moved later). RVO is assumed for the getFeatureWeights method; if
it is not RVO'd it is a small vector and only called once per solution.

This requires that the database have FeatureInfo populated with weights.
See catsop/CATMAID@8dc5b52c06ff8f3b4264ecb57a7d8d7f701aa231 to populate
CATMAID with the default weights.