Allow forced explanation of maximal cliques
SolutionGuarantor may force explanation of maximal cliques, i.e.,
require that one slice in the clique is chosen by the solver.
ConflictSets now have a boolean to indicate whether they are maximal
cliques, which is false by default but true for MSER paths converted to
ConflictSets by ComponentTreeConverter. While in the general case these
constraints over maximal cliques of the conflict graph do not force
maximally explaining solutions, they are sufficient for graphs where
each maximal clique contains at least one unique slice (of which
component trees are a subset).

Force explanation also forces any slices not in a ConflictSet to be in
the solution.

Since currently only pairwise conflicts (edges in the conflict graph)
are stored in PostgreSQL, all ConflictSets retrieved from that store are
marked as not being maximal cliques.