Allow stores to cache costs
Allow stores to save and retrieve segment costs, which are now a
property of SegmentDescriptions. This allows costs to be cached when
feature weights are unchanged. It is the responsibility of the store to
know whether costs have been invalidated (since it also stores the
weights). Greatly increases the speed of re-solution, mostly due to
needing to retrieve less information from the SegmentStore.