PostgreSqlSliceStore: fix conflict set edge case
Fix an edge case when storing conflict sets where all conflict sets were
only self-conflicts. Self-conflicts are not stored in PostgreSQL, since
conflicts are graph edges and all slices are assumed to be in conflict
with themselves. This resulted in an invalid and pointless insert.
Instead do nothing in this case (since there is no conflicts flag to
set).