SolutionGuarantor: fix conflicting segments in solution
Fix a bug where segments containing conflicting slices were generated
because the conflict set was not associated to blocks in the request.
The approach of this solution is to retrieve conflict sets for the set
of blocks containing the segments in the request blocks, rather than
only retrieving conflict sets for the request blocks. Other potential
solutions are discussed below. This is a far-reaching issue, so the bug
is described in detail since it and this solution impact future design
decisions and will likely be revisited.

The general cause of this bug, conflict sets relevant to a solution not
being associated to the set of request blocks, has two origins. The
first is that segments on either z-extent of the request blocks (core
plus padding) involve slices in sections in neighboring z-blocks. This
is the source of the vast majority of confliciting segments in
solutions, and could be fixed easily by expanding the request blocks by
1 block in z in each direction. The second cause, requiring more
elaborate solutions, involves the separation between slice and segment
guarantors. Slice guarantors associate only slices (and conflict sets)
involved with a slice overlapping each block's bounds to that block.
However, the segment guarantor generates segments for a block involving
slices within the section overlap parameter for the set of slices
associated to a block. Thus a block's associated segments involve slices
(and conflict sets) not associated to that block. In practice
conflicting solutions caused by these transitive slice associations are
exceedingly rare (< 0.01% of solution slices in L3), but do occur. More
critically, their prevalence depends on tissue structure.

One solution would be to eliminate conflict set to block relationships
completely and retrieve conflict sets directly by a set of slices.
While parsimonious, querying conflict sets this way quickly becomes
infeasible for large volumes as it is not block-indexed.

Another solution would be for SegmentGuarantor to associate conflict
sets for the slices of the segments it associates to blocks to those
same blocks. While simplying the retrieval of conflict sets for
SolutionGuarantor, this complicates the conflict set to block
relationship by making it the only store relationship to be modified by
two guarantors. It also would make SegmentGuarantor responsible for
storing information it did not generate. Going forward it may be a more
scalable solution that the one used here, but this one was chosen for
the sake of transparency, simplicity, and single responsibility.