SolutionGuarantor: handle explanation of leaf slices
Force explanation did not work for many cases because of model
infeasibility due to overlap and continuation constraints on leaf
slices. Leaf slices are those slices involved in a segment associated to
the request blocks but not directly associated to the request blocks.
Consequently end segments for these slices are not associated to the
blocks and are not in the solution, nor should they be, since they
indicate continuation to adjoining blocks. First and last slices are
special, trivial cases of leaf slices.

Do not force explanation of conflict sets involving leaf slices. Since
some leaf slices may have segments on only one side, do not enforce
continuation constraints in those cases.

For clarity the code still handles first and last slices separately from
the rest of leaf slices, but the logic may be combined in the future.