SolutionGuarantor: merge first/last slice logic with leaf slices
The only functional change of this merge is that first/last slices not
in a conflict set are no longer forced to be explained. Since these
slices are in the padding, not the core, this behavior is acceptable
(except for at the volume boundaries, where explanation is still forced
because they are not leaves).