PostgreSqlSliceStore: set slices flag if no slices
If done with a block, still set the slices flag even if no slices are
associated to the block. In general this occurs for blocks outside
imaging/prediction boundaries. In the future there may be a separate
semantic flag for this condition, but for now the slices flag must be
set to allow guarantor's block expansion to work correctly.