SolutionGuarantor: only store solution for core
Discard segments in the solution that are only in the padding blocks and
do not intersect the core. This was the planned behavior, since the
padding only provides context for the solution to each core. Culling the
solution also reduces the time and space cost of solution storage and
and assembly operations.