SolutionGuarantor: extract assemblies for solution
Because the solution guarantor already has segments loaded in memory, it
is efficient to compute connected components (assemblies) as part of the
solution.

While NeuronExtractor from upstream sopnet (removed in cc96688) could
not be reused here because of differing data structures, the approach is
similar to the current implementation in funkey/sopnet@1ff2079.

Note that the assemblies are discarded for now, but will be stored in a
future SegmentStore refactor.