PostgreSqlSliceStore: fix component loading segfault
The recent merge of upstream SOPNET imageprocessing submdoule caused a
segfault when loading connected component images, because PixelList
expects to be created with capacity sufficient to hold all pixels added
to it. Count nonzero pixels in the bitmap after loading the image so
that PixelList can be initialized properly. Also use this count in place
of the image min/max for working around the vigra single-intensity-image
bug.