PostgreSqlSegmentStore: improve segment cost locking
When storing solution costs, locking is necessary to prevent row-level
deadlocks, because PostgreSQL will update rows in arbitrary order.
Instead of locking the entire segments table, lock only the rows to be
updated. This is possible because SELECT .. FOR UPDATE respects row
order, so as long as all connections attempt to lock rows in this way
with the same order there should be no deadlocks. Make use of a
temporary table, like other store methods, to reduce transmitted data
and work done during the lock.