Add binary to produce label images for single, local core
Add a binary that produces a label image stack (with labels as the
assembly hash) from a membrane probability image stack, a raw image
stack and a set of feature weights. This is intended as a minimal
demonstration of segmentation without requiring a persistence provider,
so local stores are used.  This requires that all extracted data fit in
memory. Slice and segment extraction can be performed blockwise, but
solution is limited to a single core to produce univocal labels.