PostgreSqlSliceStore: store component masks in database
Fix issues of locality of slice component mask storage and filesystem
inode limits by storing component PNGs in the database. This
implementation has similar performance to the filesystem storage.

Because Vigra does not support loading images from memory, PNGs from the
database must be written to a temporary location on the filesystem. On
Linux this can be /dev/shm -- shared memory -- so that this limitation
has minimal performance impact.