SegmentGuarantor: expand to fewer blocks
When expanding blocks to find other slices that may have segments with
slices in the request blocks, be more efficient in the blocks selected
for expansion. Before, blocks intersecting the axis aligned bounding box
of the axis aligned bounding boxes of all slices in the request blocks
were used. In many pathological cases, such as large, diagonal slices,
this could result in expanding to many blocks that were not near the
request block slices.

Instead, expand to the union of blocks intersecting each slice's axis
aligned bounding box. This is still vulnerable to pathological cases,
but is at subset of the prior approach for many cases.