Persistence: support stack type for slices/segments
Slices and segments are associated with a stack type. For example,
during normal solving they are associated with the membrane stack, but
during ground truth extraction they are associated with that stack.
Slices and segments from ground truth should not be stored with those
from the membrane stack, since that would confuse the gold standard.

Instead, BackendClient now accepts a stack type when creating all
stores, not just stack stores, and PostgreSQL stores use the stack type
to select a schema.